Split ReceiverController: extract pair-flow
Sixth in the receiver-split arc. Extracts record_pair_request,
lookup_peer_for_session, lookup_peer_for_status, and
fire_pair_status_changed plus the shared _lookup_peer_response
helper into pair_flow.py. The four public entry points stay
as bound delegates on the controller because the peer-link
Noise WS dispatcher calls them via the controller surface;
_fire_pair_status_changed also stays as a delegate because
peer_crud calls it via the controller.

Receiver: 649 -> 495 lines (-154). Cumulative arc: 1083 ->
495 (-588, -54%). All 3418 tests pass.
